Make the cover into a container
