+++ osu.Game.Rulesets.Mania/Mods/ManiaModFadeIn.cs
@@ -3,9 +3,7 @@
 
 using osu.Framework.Graphics.Sprites;
 using osu.Game.Graphics;
-using osu.Game.Rulesets.Mania.Objects;
-using osu.Game.Rulesets.UI;
-using osuTK;
+using osu.Game.Rulesets.Mania.UI;
 
 namespace osu.Game.Rulesets.Mania.Mods
 {
@@ -16,10 +14,6 @@ namespace osu.Game.Rulesets.Mania.Mods
         public override IconUsage? Icon => OsuIcon.ModHidden;
         public override string Description => @"Keys appear out of nowhere!";
 
-        public override void ApplyToDrawableRuleset(DrawableRuleset<ManiaHitObject> drawableRuleset)
-        {
-            base.ApplyToDrawableRuleset(drawableRuleset);
-            LaneCovers.ForEach(laneCover => laneCover.Scale = new Vector2(1f, -1f));
-        }
+        protected override PlayfieldCoveringContainer CreateCover() => new PlayfieldCoveringContainer();
     }
 }
